# packages/contract/openapi.yaml
openapi: 3.1.0
info:
  title: RMC ERP API
  version: 0.2.0
  description: >
    Contract-first source of truth. The TS client, React Query hooks, and
    Zod request/response schemas consumed by apps/web and apps/mobile are
    generated from this file — see CLAUDE.md. Do not hand-write a fetch
    call that bypasses the generated hooks.

    Money fields are JOD decimal strings (e.g. "1234.567"), never numbers
    — formatting/parsing to the internal integer-fils representation
    happens at this HTTP boundary (packages/core's jodStringToFils /
    filsToJodString), matching CLAUDE.md's money conventions.
servers:
  - url: /api
...
paths:
...
  /company:
    get:
      operationId: getCompany
      tags: [company]
      summary: The single tenant company row.
      security: [{ bearerAuth: [] }]
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Company" }
        "401": { $ref: "#/components/responses/Unauthorized" }
    put:
      operationId: updateCompany
      tags: [company]
      summary: Update company settings.
      security: [{ bearerAuth: [] }]
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/UpdateCompanyRequest" }
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Company" }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "403": { $ref: "#/components/responses/Forbidden" }
        "400": { $ref: "#/components/responses/ValidationError" }
  /branches:
    get:
      operationId: listBranches
      tags: [branches]
      summary: List branches.
      security: [{ bearerAuth: [] }]
      parameters:
        - $ref: "#/components/parameters/PageParam"
        - $ref: "#/components/parameters/PageSizeParam"
        - $ref: "#/components/parameters/QParam"
        - $ref: "#/components/parameters/FormatParam"
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema: { $ref: "#/components/schemas/BranchListResponse" }
            text/csv:
              schema: { type: string }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "403": { $ref: "#/components/responses/Forbidden" }
    post:
      operationId: createBranch
      tags: [branches]
      summary: Create a branch.
      security: [{ bearerAuth: [] }]
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/CreateBranchRequest" }
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Branch" }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "403": { $ref: "#/components/responses/Forbidden" }
        "400": { $ref: "#/components/responses/ValidationError" }
  /branches/{id}:
    parameters:
      - $ref: "#/components/parameters/IdParam"
    get:
      operationId: getBranch
      tags: [branches]
      summary: Get a branch by id.
      security: [{ bearerAuth: [] }]
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Branch" }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "403": { $ref: "#/components/responses/Forbidden" }
        "404": { $ref: "#/components/responses/NotFound" }
    put:
      operationId: updateBranch
      tags: [branches]
      summary: Update a branch.
      security: [{ bearerAuth: [] }]
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/UpdateBranchRequest" }
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Branch" }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "403": { $ref: "#/components/responses/Forbidden" }
        "404": { $ref: "#/components/responses/NotFound" }
        "400": { $ref: "#/components/responses/ValidationError" }
    delete:
      operationId: voidBranch
      tags: [branches]
      summary: Soft-delete (void) a branch.
      security: [{ bearerAuth: [] }]
      requestBody:
        content:
          application/json:
            schema: { $ref: "#/components/schemas/VoidRequest" }
      responses:
        "204": { description: Voided }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "403": { $ref: "#/components/responses/Forbidden" }
        "404": { $ref: "#/components/responses/NotFound" }
...
  /customers:
    get:
      operationId: listCustomers
      tags: [customers]
      summary: List customers.
      security: [{ bearerAuth: [] }]
      parameters:
        - $ref: "#/components/parameters/PageParam"
        - $ref: "#/components/parameters/PageSizeParam"
        - $ref: "#/components/parameters/QParam"
        - $ref: "#/components/parameters/FormatParam"
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema: { $ref: "#/components/schemas/CustomerListResponse" }
            text/csv:
              schema: { type: string }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "403": { $ref: "#/components/responses/Forbidden" }
    post:
      operationId: createCustomer
      tags: [customers]
      summary: Create a customer.
      security: [{ bearerAuth: [] }]
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/CreateCustomerRequest" }
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Customer" }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "403": { $ref: "#/components/responses/Forbidden" }
        "400": { $ref: "#/components/responses/ValidationError" }
  /customers/{id}:
    parameters:
      - $ref: "#/components/parameters/IdParam"
    get:
      operationId: getCustomer
      tags: [customers]
      summary: Get a customer by id.
      security: [{ bearerAuth: [] }]
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Customer" }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "403": { $ref: "#/components/responses/Forbidden" }
        "404": { $ref: "#/components/responses/NotFound" }
    put:
      operationId: updateCustomer
      tags: [customers]
      summary: Update a customer.
      security: [{ bearerAuth: [] }]
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/UpdateCustomerRequest" }
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Customer" }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "403": { $ref: "#/components/responses/Forbidden" }
        "404": { $ref: "#/components/responses/NotFound" }
        "400": { $ref: "#/components/responses/ValidationError" }
    delete:
      operationId: voidCustomer
      tags: [customers]
      summary: Soft-delete (void) a customer.
      security: [{ bearerAuth: [] }]
      requestBody:
        content:
          application/json:
            schema: { $ref: "#/components/schemas/VoidRequest" }
      responses:
        "204": { description: Voided }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "403": { $ref: "#/components/responses/Forbidden" }
        "404": { $ref: "#/components/responses/NotFound" }
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  parameters:
    IdParam:
      name: id
      in: path
      required: true
      schema: { type: string, format: uuid }
    PageParam:
      name: page
      in: query
      schema: { type: integer, minimum: 1 }
    PageSizeParam:
      name: pageSize
      in: query
      schema: { type: integer, minimum: 1, maximum: 100 }
    QParam:
      name: q
      in: query
      description: Free-text search.
      schema: { type: string }
    FormatParam:
      name: format
      in: query
      schema: { type: string, enum: [json, csv] }
  responses:
    Unauthorized:
      description: Not authenticated
      content:
        application/json:
          schema: { $ref: "#/components/schemas/ErrorResponse" }
    Forbidden:
      description: Missing permission
      content:
        application/json:
          schema: { $ref: "#/components/schemas/ErrorResponse" }
    NotFound:
      description: Not found
      content:
        application/json:
          schema: { $ref: "#/components/schemas/ErrorResponse" }
    ValidationError:
      description: Validation error
      content:
        application/json:
          schema: { $ref: "#/components/schemas/ErrorResponse" }
  schemas:
...
    ErrorResponse:
      type: object
      required: [error]
      properties:
        error:
          type: object
          required: [message, code]
          properties:
            message: { type: string }
            code: { type: string }
            details: { type: object }
    VoidRequest:
      type: object
      properties:
        reason: { type: string }
    Company:
      type: object
      required: [id, name, taxNumber, createdAt, updatedAt]
      properties:
        id: { type: string, format: uuid }
        name: { type: string }
        taxNumber: { type: string, nullable: true }
        createdAt: { type: string, format: date-time }
        updatedAt: { type: string, format: date-time }
    UpdateCompanyRequest:
      type: object
      properties:
        name: { type: string, minLength: 1 }
        taxNumber: { type: string, nullable: true }
    Branch:
      type: object
      required: [id, companyId, name, code, createdAt, updatedAt, voidedAt]
      properties:
        id: { type: string, format: uuid }
        companyId: { type: string, format: uuid }
        name: { type: string }
        code: { type: string }
        createdAt: { type: string, format: date-time }
        updatedAt: { type: string, format: date-time }
        voidedAt: { type: string, format: date-time, nullable: true }
    CreateBranchRequest:
      type: object
      required: [name, code]
      properties:
        name: { type: string, minLength: 1 }
        code: { type: string, minLength: 1, maxLength: 32 }
    UpdateBranchRequest:
      type: object
      properties:
        name: { type: string, minLength: 1 }
        code: { type: string, minLength: 1, maxLength: 32 }
    BranchListResponse:
      type: object
      required: [items, total, page, pageSize]
      properties:
        items: { type: array, items: { $ref: "#/components/schemas/Branch" } }
        total: { type: integer }
        page: { type: integer }
        pageSize: { type: integer }
...
    Customer:
      type: object
      required:
        [id, companyId, branchId, name, customerType, phone, email, taxNumber, nationalId, address,
         creditLimitJod, creditPolicy, paymentTermsDays, notes, createdAt, updatedAt, voidedAt]
      properties:
        id: { type: string, format: uuid }
        companyId: { type: string, format: uuid }
        branchId: { type: string, format: uuid, nullable: true }
        name: { type: string }
        customerType: { type: string, enum: [individual, company] }
        phone: { type: string, nullable: true }
        email: { type: string, nullable: true }
        taxNumber: { type: string, nullable: true }
        nationalId: { type: string, nullable: true }
        address: { type: string, nullable: true }
        creditLimitJod: { type: string, description: "Decimal JOD, e.g. \"5000.000\"" }
        creditPolicy: { type: string, enum: [none, warning, block] }
        paymentTermsDays: { type: integer }
        notes: { type: string, nullable: true }
        createdAt: { type: string, format: date-time }
        updatedAt: { type: string, format: date-time }
        voidedAt: { type: string, format: date-time, nullable: true }
    CreateCustomerRequest:
      type: object
      required: [name]
      properties:
        name: { type: string, minLength: 1 }
        customerType: { type: string, enum: [individual, company] }
        branchId: { type: string, format: uuid, nullable: true }
        phone: { type: string, nullable: true }
        email: { type: string, nullable: true }
        taxNumber: { type: string, nullable: true }
        nationalId: { type: string, nullable: true }
        address: { type: string, nullable: true }
        creditLimitJod: { type: string }
        creditPolicy: { type: string, enum: [none, warning, block] }
        paymentTermsDays: { type: integer }
        notes: { type: string, nullable: true }
    UpdateCustomerRequest:
      type: object
      properties:
        name: { type: string, minLength: 1 }
        customerType: { type: string, enum: [individual, company] }
        branchId: { type: string, format: uuid, nullable: true }
        phone: { type: string, nullable: true }
        email: { type: string, nullable: true }
        taxNumber: { type: string, nullable: true }
        nationalId: { type: string, nullable: true }
        address: { type: string, nullable: true }
        creditLimitJod: { type: string }
        creditPolicy: { type: string, enum: [none, warning, block] }
        paymentTermsDays: { type: integer }
        notes: { type: string, nullable: true }
    CustomerListResponse:
      type: object
      required: [items, total, page, pageSize]
      properties:
        items: { type: array, items: { $ref: "#/components/schemas/Customer" } }
        total: { type: integer }
        page: { type: integer }
        pageSize: { type: integer }
...